RcppEigen-package {RcppEigen}	R Documentation

Rcpp/Eigen bridge

Description

The package eases the use of the Eigen C++ template library for linear algebra with Rcpp

Details

This package contains the header files for the Eigen C++ template library. The typical usage is to install this package and list it in the LinkingTo: line in the ‘DESCRIPTION’ file of other packages. The C++ source code and the R source code in this package are for illustration only.

As described at the Eigen project's home page, http://eigen.tuxfamily.org/index.php?title=Main_Page, Eigen is a versatile, fast, reliable and elegant collection of C++ classes for linear algebra.

References

Douglas Bates and Dirk Eddelbuettel (2013). Fast and Elegant Numerical Linear Algebra Using the RcppEigen Package. Journal of Statistical Software, 52(5), 1-24. URL http://www.jstatsoft.org/v52/i05/.

[Package RcppEigen version 0.3.3.9.2 Index]